Move ListItem components to separate folders
