+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/list-item/components/index.ts
@@ -1,7 +1,7 @@
-export * from './ListItemContent';
-export * from './ListItemItem';
-export * from './ListItemGroup';
-export * from './ListItemLabel';
-export * from './ListItemText';
-export * from './ListItemTrigger';
-export * from './ListItemFooter';
+export * from './list-item-content';
+export * from './list-item-item';
+export * from './list-item-group';
+export * from './list-item-label';
+export * from './list-item-text';
+export * from './list-item-trigger';
+export * from './list-item-footer';
