+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/list-item/components/list-item-content/index.ts
@@ -1,0 +1,1 @@
+export * from './ListItemContent';
