+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/list-item/components/list-item-group/ListItemGroup.tsx
@@ -1,4 +1,4 @@
-import { Flex, FlexProps } from '../../flex';
+import { Flex, FlexProps } from '../../../flex';
 
 export type ListItemGroupProps = FlexProps;
