+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/list-item/components/list-item-item/hooks/useBackgroundColor.tsx
@@ -1,0 +1,7 @@
+import { levels } from '../../../levels';
+import { useListItem } from '../../../ListItemContext';
+
+export const useBackgroundColor = () => {
+  const { level, disabled } = useListItem();
+  return disabled ? levels[level].disabled : levels[level].enabled;
+};
