+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/list-item/components/list-item-label/ListItemLabel.tsx
@@ -1,6 +1,6 @@
-import { Colors } from '../../../foundations';
-import { LabelTinyProps, TitleMedium } from '../../typography';
-import { useListItem } from '../ListItemContext';
+import { Colors } from '../../../../foundations';
+import { LabelTinyProps, TitleMedium } from '../../../typography';
+import { useListItem } from '../../ListItemContext';
 
 export type ListItemLabelProps<E extends React.ElementType = 'span'> = LabelTinyProps<E>;
