+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/list-item/components/list-item-text/index.ts
@@ -1,0 +1,1 @@
+export * from './ListItemText';
